Refactor blog utilities: move formatting functions to blog-format module
