+++ src/components/home/HomeSocialProofSection.tsx
@@ -5,7 +5,7 @@ import { ArrowRight } from 'lucide-react'
 import { Card } from '~/components/Card'
 import { PartnersGrid } from '~/components/PartnersGrid'
 import { Button } from '~/ui'
-import { formatAuthors, formatPublishedDate } from '~/utils/blog'
+import { formatAuthors, formatPublishedDate } from '~/utils/blog-format'
 import type { RecentPost } from '~/utils/blog.functions'
 
 type HomeSocialProofSectionProps = {
