+++ src/routes/rss[.]xml.ts
@@ -1,10 +1,10 @@
 import { createFileRoute } from '@tanstack/react-router'
 import { setResponseHeader } from '@tanstack/react-start/server'
+import { getPublishedPosts } from '~/utils/blog'
 import {
-  getPublishedPosts,
   formatAuthors,
   publishedDateToUTCString,
-} from '~/utils/blog'
+} from '~/utils/blog-format'
 
 function escapeXml(unsafe: string): string {
   return unsafe
